Move transitive reference analyzer to step after other analyzers
The transitive dependency analyzer often seems to cause problems with loops. This change moves it so that it occurs after all the analyzers are run by moving it into its own step.
