+++ src/extensions/default/Microsoft.DotNet.UpgradeAssistant.Steps.Packages/PackageUpdaterStepExtensions.cs
@@ -23,6 +23,7 @@ namespace Microsoft.DotNet.UpgradeAssistant
 
             services.Services.AddUpgradeStep<PackageUpdaterPreTFMStep>();
             services.Services.AddUpgradeStep<PackageUpdaterStep>();
+            services.Services.AddUpgradeStep<TransitiveReferenceUpgradeStep>();
             services.Services.AddTransient<IDependencyAnalyzerRunner, DependencyAnalyzerRunner>();
             services.Services.AddTransient<IAnalyzeResultProvider, PackageAnalyzerResultProvider>();
